--- src/main/java/LoadBalancer/LBStrategy.java
@@ -1,9 +1,0 @@
-package LoadBalancer;
-
-public abstract class LBStrategy {
-
-
-    public LBStrategy(String fleetManager) {}
-
-    public abstract String getNextHost();
-}
