+++ src/main/java/requestManagement/loadBalancer/LBStrategy.java
@@ -1,0 +1,11 @@
+package requestManagement.loadBalancer;
+
+import requestManagement.fleetManager.FleetManager;
+
+public abstract class LBStrategy {
+
+
+    public LBStrategy(FleetManager fleetManager) {}
+
+    public abstract String getNextHost();
+}
