--- src/main/java/requestManagement/loadBalancing/LoadBalancer.java
@@ -1,3 +1,0 @@
-package requestManagement.loadBalancing;
-
-public class LoadBalancer {}
