+++ src/main/test/LoadBalancerTest.java
@@ -1,12 +1,15 @@
-import LoadBalancer.LBRoundRobin;
-import LoadBalancer.LBStrategy;
-import LoadBalancer.LoadBalancer;
+import requestManagement.fleetManager.FleetManager;
+import requestManagement.loadBalancer.LBRoundRobin;
+import requestManagement.loadBalancer.LBStrategy;
+import requestManagement.loadBalancer.LoadBalancer;
 import org.junit.jupiter.api.Test;
+
+import static org.mockito.Mockito.mock;
 
 public class LoadBalancerTest {
     @Test
     void loadBalancerBuilderTest() {
-        String fleetManager = new String();
+        FleetManager fleetManager = mock(FleetManager.class);
         LBStrategy loadBalancingStrategy = new LBRoundRobin(fleetManager);
 
         LoadBalancer loadBalancer = LoadBalancer.getBuilder()
